separate VariantSearchFormPanels
